Initial version of the scatter-gather filesystem.
This is a filesystem that uses multiple other filesystems as backend storage.
Contrary to "unionfs", "overlayfs" and similar stacking filesystems, a file
with a given path can only be present in one of the backend filesystems, and
the backends are either all read-only or all read-write.  The filesystem will
appear as the union of all backends.  File operations are passed through to the
file on the backend that provides that file.  For new files, heuristics will
determine to which of the backend filesystems they will be written to; it can
either try to fill up one backend before moving to the next, spread out files
as evenly as possible, or try to maintain locality (based on the path).

The intended use is to create huge filesystems from multiple independent
filesystems, where adding storage is completely trivial (no need to grow
filesystems or RAID sets), and graceful degredation in case of problems with
one of the backend filesystems. Backend filesystems can still be used
independently when needed.

The need for this filesystem arose from having to deal with huge filesystems
where the effective downtime caused by RAID resyncs, fsck, growfs, etc. is also
huge. Using multiple, smaller RAID sets and filesystems helps a lot, but one
loses the benefit of having a single mountpoint where you can just write files
to until everything is full. This filesystem provides the latter functionality.

This filesystem does not do striping/mirroring/redundancy in any way. The
underlying filesystems should provide that if necessary. Growing the SGFS is
easy, just remount it with more backends. Shrinking is also relatively trivial;
remount with one of the backends removed, and copy/move everything from the
removed backend into the SGFS, which will ensure the files from the removed
backend are now available from one or more of the remaining backends.

At the moment, SGFS is implemented using FUSE, with enough functionality to
read/write/create files.