Fix (sym)linking.
For sgfs_symlink(), from is just the data put into the symlink, while to is a
real path.

For sgfs_link(), we try to create the hardlink on the same underlay as the file
referenced. This version does not try to create the necessary parent
directories if they do not exist yet on that underlay though.