Use open(O_CREAT | O_EXCL) instead of mknod()ing regular files.
We cache the filedescriptor for subsequent sgfs_open() calls. Apart from saving
some syscalls, this is necessary to handle someone opening a new file without
write permissions (for example, mode 0400) with O_RDWR, and writing to it. Yes,
this is normal behaviour, and without this, copying a git repository to a SGFS
doesn't work.