Much improved sgfs_rename().
We make a distinction between renaming directories and renaming files.  In case
of directories, we check all underlays for the original directory name, and if
it exists, rename it in that underlay.  For files, we just rename in the first
encounter, since we assume there it is unique.

In each case, before we do a renameat(), we fix the directory tree in that
underlay to ensure the rename can happen.

Still broken is renaming a file that is in one underlay to a file that exists
in another underlay.